requirements.txt
Flask==0.10.1
Flask-Redis==0.1.0
Flask-WTF==0.12
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
python-Levenshtein==0.12.0
redis==2.10.5
requests==2.9.1
Werkzeug==0.11.4
WTForms==2.1
